Move card components to subfolder
So that we can begin to build simplified components which store and
manage their own state, rather than relying on large and complex
component files.
